[hyperactor] mesh: concurrent spawns, aggregated statuses
This adds concurrent spawning for procs, and introduces a principled cast/accumulate status aggregation.

The latter lets us efficiently determine the status of a mesh, and apply whatever policy makes sense in context (e.g., early exit).

We will use GetRankStatus also to apply liveness checks for actors, as we can support this efficiently.

This change also introduces an RLE-encoded data structure, RankedValues, which works well for low-cardinality contexts (e.g., statuses); but to use it more generally for value meshes will require that the merge operation be optimized.

Differential Revision: [D83678884](https://our.internmc.facebook.com/intern/diff/D83678884/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D83678884/)!

ghstack-source-id: 313398128
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1394